Change css for margin and width
Changed the structure of css for the main article width and large
margin to limit the number of declarations. Also had to include new
unload script to size full bleed images. Still need to add event to do
same for resizing events. And need further cleaning of code for
responsiveness.