Fix for flaw in network.json contruction
This was originally tested with artists and themes listed in only one
post. When other posts were introduced, an issue with the nested for
loops became apparent: for post in site.posts, and then for item in
post.network. Because of this nesting I was having trouble getting it
to recognize the last edge array in the series and omitting the final
comma in that case. This is the best I cam up with. It works, but
requires that there is a network artist and theme listed on the
earliest post.